Updated the move method to reflect new location

+++ lambda/open_data_export/main.py
@@ -2,6 +2,7 @@ import logging
 import os
 import csv
 import gzip
+import re
 
 from multiprocessing import Process, Pool
 from open_data_export.pgdb import DB
@@ -206,7 +207,8 @@ def object_exists(Bucket: str, Key: str):
     return exists
 
 
-def move_object(from_location, to_location):
+def copy_object(from_location, to_location, delete=False):
+    logger.debug(f"moving {from_location['Bucket']}/{from_location['Key']}\n to {to_location['Bucket']}/{to_location['Key']}")
     try:
         if (to_location['Bucket'] != from_location['Bucket'] or to_location['Key'] != from_location['Key']):
             s3.copy_object(
@@ -219,11 +221,12 @@ def move_object(from_location, to_location):
                 },
             )
             logger.debug(f"Copied: {to_location['Bucket']}/{to_location['Key']}")
-            s3.delete_object(
-               Bucket=from_location['Bucket'],
-               Key=from_location['Key'],
-            )
-            logger.debug(f"Deleted: {from_location['Bucket']}/{from_location['Key']}")
+            if delete:
+                s3.delete_object(
+                Bucket=from_location['Bucket'],
+                Key=from_location['Key'],
+                )
+                logger.debug(f"Deleted: {from_location['Bucket']}/{from_location['Key']}")
         else:
             # if its the same lets just update the ACL policy
             s3.put_object_acl(
@@ -254,7 +257,7 @@ def move_object(from_location, to_location):
 def move_objects_handler(event, context=None):
     start = time.time()
     db = get_database()
-    limit = 5000
+    limit = 2
     where = "exported_on IS NOT NULL"
     # AND (l.metadata->>'move' IS NOT NULL AND (l.metadata->>'move')::boolean = true)
     args = {}
@@ -270,6 +273,15 @@ def move_objects_handler(event, context=None):
         args['node'] = event['node']
         where += " AND l.sensor_nodes_id = :node"
 
+    where = """
+			COALESCE(key, l.metadata->>'Key') IS NOT NULL
+			AND COALESCE(key, l.metadata->>'Key') ~* :pattern
+			AND l.exported_on IS NOT NULL
+			AND l.records > :records
+			"""
+
+    args = { "records": 0, "pattern": "/provider" }
+
     # determine the extension type
     if settings.WRITE_FILE_FORMAT == 'csv':
         ext = 'csv'
@@ -284,50 +296,41 @@ def move_objects_handler(event, context=None):
 
     # where = " AND l.metadata->>'Bucket' IS NOT NULL"
 
-    days = db.rows(
+    days, time_ms = db.rows(
         f"""
         WITH days AS (
         -----------
         -- get a set of files to move
         -----------
-        SELECT
+         SELECT
           l.day
         , l.sensor_nodes_id
-        , lower(COALESCE(sn.country, 'no-country')) as country
-        , p.export_prefix
-        , l.open_data_export_logs_id
-        , FORMAT('records/{ext}/provider=%%s/country=%%s/locationid=%%s/year=%%s/month=%%s/location-%%s-%%s.{ext}'
-          , p.export_prefix
-          , lower(COALESCE(sn.country, 'no-country'))
+        , COALESCE(l.key, l.metadata->>'Key') as from_key
+        , FORMAT('records/csv.gz/country=%%s/locationid=%%s/year=%%s/month=%%s/location-%%s-%%s.{ext}'
+          , lower(COALESCE(c.iso, 'no-country'))
           , l.sensor_nodes_id
           , to_char(l.day, 'YYYY')
           , to_char(l.day, 'MM')
           , l.sensor_nodes_id
           , to_char(l.day, 'YYYYMMDD')
-        ) as key
-        , l.metadata->>'Key' as current_key
-        , l.metadata->>'error' as current_error
-        , l.metadata->>'message' as current_error_message
+        ) as to_key
         FROM open_data_export_logs l
         JOIN sensor_nodes sn ON (l.sensor_nodes_id = sn.sensor_nodes_id)
-        JOIN providers p ON (sn.source_name = p.source_name)
+        JOIN countries c ON (sn.countries_id = c.countries_id)
         WHERE {where}
         LIMIT {limit})
         -----------
         -- Update those records and return data
         -----------
         UPDATE open_data_export_logs
-        SET metadata = COALESCE(metadata::jsonb, '{{}}'::jsonb)
-        ||jsonb_build_object(
-           'Bucket', '{settings.OPEN_DATA_BUCKET}'
-          ,'Key', key
-          ,'move', false
-          ,'error', false
-          ,'message', 'no error'
+        SET key = 's3://{settings.OPEN_DATA_BUCKET}/'||to_key
+          , metadata = jsonb_build_object(
+           'moved_on', now(),
+           'moved_from', from_key
         )
         FROM days
-        WHERE days.open_data_export_logs_id =
-              open_data_export_logs.open_data_export_logs_id
+        WHERE days.sensor_nodes_id=open_data_export_logs.sensor_nodes_id
+        AND days.day=open_data_export_logs.day
         -----------
         -- return the pre-update data
         -----------
@@ -335,41 +338,40 @@ def move_objects_handler(event, context=None):
         """, **args)
 
     successes = 0
+    regex = re.compile('s3://[a-z-]+/', re.IGNORECASE)
     for row in days:
-
-        node = row[1]
-        country = row[2]
+        logger.debug(row)
         day = row[0]
-        key = row[5]
-        old_key = row[6]
+        node = row[1]
+        from_key = row[2]
+        to_key = row[3]
 
         yr = day.strftime('%Y')
         mn = day.strftime('%m')
         dy = day.strftime('%d')
 
-        if row[7] is not None:
-            logger.debug(f"Previous error: {row[8]}")
-
-        # if old_key is None:
-        old_key = f"""records/{settings.WRITE_FILE_FORMAT}/country={country}/locationid={node}/year={yr}/month={mn}/location-{node}-{yr}{mn}{dy}.{ext}"""
-
         try:
-            move_object(
+            # make sure that we have the extension on the from_key
+            if not from_key.endswith(ext):
+                from_key = f"{from_key}.{ext}"
+
+            copy_object(
                 from_location={
                     "Bucket": settings.OPEN_DATA_BUCKET,
-                    "Key": old_key,
+                    "Key": regex.sub('', from_key),
                 },
                 to_location={
                     "Bucket": settings.OPEN_DATA_BUCKET,
-                    "Key": key,
-                }
+                    "Key": to_key,
+                },
+				delete=False
             )
             successes += 1
         except Exception as e:
-            submit_error(day, node, f"{e}")
+            submit_move_error(day, node, from_key, f"{e}")
 
     sec = time.time() - start
-    logger.info(f'Moved {successes} files (of {len(days)}) in {sec} seconds')
+    logger.info(f'Moved {successes} files (of {len(days)}) in {sec} seconds (query: {time_ms/1000})')
 
 
 def update_hourly_data(ts):
@@ -449,6 +451,27 @@ def submit_error(day: str, node: int, error: str):
     logger.error(f"error: {node} on {day} - {error}")
     db = get_database()
     return db.rows(sql, day=day, node=node, error=error)
+
+def submit_move_error(day: str, node: int, key: str, error: str):
+    """
+    Mark the location/day with an error message and revert the key
+    """
+    if isinstance(day, str):
+        day = datetime.fromisoformat(day).date()
+    sql = """
+    UPDATE open_data_export_logs
+    SET key = :key
+	, metadata = (COALESCE(metadata::jsonb, '{}'::jsonb)||jsonb_build_object(
+      'error', true
+    , 'message', (:error)::text
+    , 'at', current_timestamp::text
+    ))::json
+    WHERE day = :day AND sensor_nodes_id = :node
+    RETURNING TRUE
+    """
+    logger.error(f"error: {node} on {day} - {error}")
+    db = get_database()
+    return db.rows(sql, day=day, node=node, key=key, error=error)
 
 
 def get_all_location_days():
@@ -751,7 +774,10 @@ def export_data(day, node):
 
 def export_data_mp(p):
     logger.info(f"Starting {p[0]}/{p[1]} on pid: {os.getpid()}")
-    n, sec = export_data(p[1], p[0])
+    try:
+        n, sec = export_data(p[1], p[0])
+    except Exception as e:
+        sec = -1
     return sec
 
 def export_pending():
@@ -783,7 +809,7 @@ def update_outdated_handler(event=None, context=None):
     if context is not None:
         time_available = context.get_remaining_time_in_millis()/1000
     else:
-        time_available = 80  # basically do it once
+        time_available = 80  # basically do it onceq
 
     time_left = time_available
     days = 0
